Container Friendliness - ENGAGE!
But plugins don't work in containers :(
