--- .dockerignore
@@ -1,5 +1,0 @@
-docker-compose.yml
-Makefile
-*.md
-.git
-.gitignore
