+++ docker-compose.yml
@@ -1,9 +1,11 @@
 plugin:
   build: .
   volumes:
-    - /run/docker/plugins/ovs.sock:/run/docker/plugins/ovs.sock
+    - /run/docker/plugins:/run/docker/plugins
     - /var/run/docker.sock:/var/run/docker.sock
   net: host
+  stdin_open: true
+  tty: true
   privileged: true
 
 ovs:
