+++ ovs/docker.go
@@ -1,0 +1,7 @@
+package ovs
+
+import "github.com/samalba/dockerclient"
+
+type dockerer struct {
+	client *dockerclient.DockerClient
+}
