--- plugin/Dockerfile
@@ -1,1 +1,0 @@
-FROM golang:1.4-onbuild
