--- plugin/ovs/docker.go
@@ -1,7 +1,0 @@
-package ovs
-
-import "github.com/samalba/dockerclient"
-
-type dockerer struct {
-	client *dockerclient.DockerClient
-}
